requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
discord.py==2.0.1
gunicorn==20.1.0
Pillow==8.4.0
python-dotenv==1.0.0
